Pellet change and buttons added
changed the pellets size so that they are more inline with real Pacman's collision. I also added a code button to bring the user to a screen where they can enter a code for an easter egg, and added an exit button to quit the program or exit back to the main menu.