Walls and other fixes
Added was that pacman can stops at and cant move through. Also added so that the pellets will not be added to the page if they are overlapping with the walls. Changed a few values and moved some things around to make it a bit more optimized and to make it look a bit cleaner than it did before