Leaderboard, power pellet, and Output Update
Added a working database with input and output with SQL. Power pellets now should work without error allowing for you to eat as many ghosts as you want (as long as their is enough power pellets to supply the hunger). Changed all System.out's to Dialog boxes that pop up making it easier to read instead of looking at the terminal for feedback.